requirements.txt
absl-py==2.1.0
aext-assistant==4.0.15
aext-assistant-server==4.0.15
aext-core==4.0.15
aext-core-server==4.0.15
aext-panels==4.0.15
aext-panels-server==4.0.15
aext-share-notebook==4.0.15
aext-share-notebook-server==4.0.15
aext-shared==4.0.15
aiobotocore==2.12.3
aiohttp==3.9.5
aioice==0.9.0
aioitertools==0.7.1
aiortc==1.9.0
aiosignal==1.2.0
alabaster==0.7.16
altair==5.0.1
anaconda-anon-usage==0.4.4
anaconda-catalogs==0.2.0
anaconda-client==1.12.3
anaconda-cloud-auth==0.5.1
anaconda-navigator==2.6.0
anaconda-project==0.11.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
archspec==0.2.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.14.2
astropy==6.1.0
astropy-iers-data==0.2024.6.3.0.31.14
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
atomicwrites==1.4.0
attrs==24.2.0
Automat==20.2.0
autopep8==2.0.4
av==12.3.0
Babel==2.11.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.4.2
bleach==4.1.0
blinker==1.6.2
bokeh==3.4.1
boltons==23.0.0
botocore==1.34.69
Bottleneck==1.3.7
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.1
conda==24.9.1
conda-build==24.5.1
conda-content-trust==0.2.0
conda-index==0.5.0
conda-libmamba-solver==24.1.0
conda-pack==0.7.1
conda-package-handling==2.3.0
conda-package-streaming==0.10.0
conda-repo-cli==1.0.88
conda-token==0.5.0+1.g2209e04
constantly==23.10.4
contourpy==1.2.0
cookiecutter==2.6.0
cryptography==42.0.5
cssselect==1.2.0
cycler==0.11.0
cytoolz==0.12.2
dask==2024.5.0
dask-expr==1.1.0
datashader==0.16.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.8
distributed==2024.5.0
distro==1.9.0
dnspython==2.7.0
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.1
flake8==7.0.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
frozendict==2.4.2
frozenlist==1.4.0
fsspec==2024.3.1
gast==0.6.0
gensim==4.3.2
gitdb==4.0.7
GitPython==3.1.37
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-speech==2.27.0
google-crc32c==1.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.11.0
HeapDict==1.0.1
holoviews==1.19.0
httpcore==1.0.2
httpx==0.27.0
hvplot==0.10.0
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
imagecodecs==2023.1.23
imageio==2.33.1
imagesize==1.4.1
imbalanced-learn==0.12.3
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.7.0
intervaltree==3.1.0
ipykernel==6.28.0
ipython==8.25.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isort==5.13.2
itemadapter==0.3.0
itemloaders==1.1.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jax==0.4.35
jaxlib==0.4.35
jedi==0.18.1
jellyfish==1.0.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==1.0.0
keras==3.6.0
keyring==24.3.1
kiwisolver==1.4.4
lazy_loader==0.4
lazy-object-proxy==1.10.0
lckr_jupyterlab_variableinspector==3.1.0
libarchive-c==2.9
libclang==18.1.1
libmambapy==1.5.8
linkify-it-py==2.0.0
llvmlite==0.42.0
lmdb==1.4.1
locket==1.0.0
lxml==5.2.1
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.0
mediapipe==0.10.14
menuinst==2.1.1
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml_dtypes==0.5.0
more-itertools==10.1.0
msgpack==1.0.5
multipledispatch==0.6.0
multitasking==0.0.11
mypy==1.10.2
nbclassic==1.0.0
nbclient==0.7.0
nbconvert==7.5.0
nbformat==5.7.3
nest-asyncio==1.5.5
networkx==3.1
notebook==6.5.5
numba==0.56.4
numpy==1.24.3
numpydoc==1.5.0
nvidia-pyindex==1.0.9
nvidia-pytorch==23.3.0
openai==0.27.4
opencv-python==4.8.0.74
opt-einsum==3.3.0
osqp==0.6.2
packaging==23.0
pandas==1.5.3
pandas-profiling==3.7.0
pandocfilters==1.5.0
param==1.12.0
parso==0.8.3
pathtools==0.1.2
pex==2.2.0
pickleshare==0.7.5
Pillow==9.5.1
pip==23.1.2
platformdirs==3.11.0
pluggy==1.0.0
protobuf==4.25.0
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.0
pycparser==2.21
pyflakes==3.2.1
pygments==2.15.1
pyjwt==2.6.0
pylint==2.17.3
pyodbc==4.0.35
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.1.0
pytest-pylint==0.23.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytesseract==0.3.10
pyyaml==6.0
quantum-gate==1.4.0
qt==5.15.9
qtconsole==5.4.0
QtPy==2.3.1
ray==2.8.0
regex==2024.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rpy2==3.5.8
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==67.4.0
sh==2.0.3
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
spacy==3.6.0
spacy-legacy==3.0.12
SQLAlchemy==2.0.18
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.13.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.2.0
toml==0.10.2
tornado==6.3.1
torch==2.1.0
torchvision==0.15.0
tqdm==4.66.1
traitlets==5.9.0
typing-extensions==4.7.1
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
vispy==0.15.3
wandb==0.15.6
wcwidth==0.2.6
webcolors==1.13
websockets==11.0.3
Werkzeug==2.3.5
xgboost==1.7.6
xlrd==2.0.1
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
